Generalize the addItem method to handle differently video, audio, image and text
Instead of using the MRL of the first subitem as the item's MRL and ignoring all rest,
now reads the "upnp:class" field of the item and the "protocolInfo" field of
each subitem to use those accordingly.
The result is that artworks contained as subitems are now used along with subtitles as slaves.

Signed-off-by: Hugo Beauzée-Luyssen <hugo@beauzee.fr>