vdpau: fix video scaling
There is a possible issue in VdpVideoMixerRender [1] call. While
destination_rect is set to dst_rect, which is fine, destination_video_rect
parameter is set to NULL, which makes video keep the size, while it's
expected to fill the entire output rectangle.

Proposed patch sets destination_video_rect to dst_rect, which makes
video scale when player window is resized.

[1] http://http.download.nvidia.com/XFree86/vdpau/doxygen/html/group___vdp_video_mixer.html#ga62bf3bf8c5f01322a03b07065c5ea3db

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>

There has been no clarifications about the intended behaviour of NULL
destination_video_rect in the VDPAU API specification; presumably the
behaviour of the NVIDIA driver is intended. But the VA-GL driver has
had a different behaviour. This patch avoids relying on either by
setting the destination_video_rect explicitly always since it is not
clear what is correct.