Win32: modify the LoadLibrary PATHS used
We used to load system libraries without the full path, notably for
DirectX-related libraries. This is a bad idea if someone puts a
similarly-named DLL in the VLC folder, because they would be loaded.

Indeed, even if we don't load from CWD, we still load from the
application, which could be an issue, if you install a DLL next to
libvlccore.dll.

Therefore, on modern Windows systems, now LoadLibrary calls are
completely limited to SYSTEM32; except when loading vlc modules,
where they are limited to the application folder.