input: fix skipping data in stream-oriented access
Now that stream cache and stream/access wrapper are separated the later
must handle skipping data (by "reading into" a NULL buffer). This was
done correctly for block-oriented plugins, but not stream-oriented
ones. This patch simply adds a dummy buffer to read into in the
unlikely event that it is needed.

(Typically one of the stream cache plugins takes care of this, so this
bug was mostly invisible.)

Pointed-out-by: Filip Roséen <filip@videolabs.io>