http: do not send GET request at end (refs #17159)
If the size of the representation of the entity is known (via the
Content-Range response header field), and the current offset is equal
or larger to that size, do not send a GET request. The response would
normally be an HTTP 416 error, unless the representation changed, in
which case it would be an HTTP 412 error. There is no point sending a
request if we know it will always fail.

Note that a request is still send at end if the server did not supply
a Content-Range header (typically because the representation size is
not known a priori). This is a bug.