access/sftp: fix argument to vlc_UrlParse
Given that p_access->psz_location is the input-mrl with the scheme part
stripped, and that part is required for vlc_UrlParse to properly
identify the host we would never be able to open up a remote resource
using the module.

This fixes the issue by instead passing p_access->psz_url.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>