gui/qt: remove premature return in PLModel::isSupportedAction
The index passed will be considered invalid if the user right-clicks on
an area that does not belong to an entity.

Given that there are options that certainly are supported when such
right-click is in flight, we should not prematurely return from the
function; instead each case should respect whether we have a non-root
item tied to the on-going request or not.

Signed-off-by: Francois Cartegnie <fcvlcdev@free.fr>