gui/qt: fix PLModel::createNode when invoked with rootIndex()
This patch fixes ticket #10051, where it is described that one cannot
create a new directory in the playlist/media-library unless one does it
by right-clicking on an existing item.

In other words; if a request to create a directory is initiated in a
context where the list is empty, it would not work.

The reason is imply because the function prematurely returns if the
passed index does not refer to a specific item, and as such the relevant
check (and premature return) is removed in this commit.

PLModel::createNode already has functionality in place to handle
creating directories as top-level of the relevant playlist, so the
premature return was just blocking this from kicking in.

Signed-off-by: Francois Cartegnie <fcvlcdev@free.fr>