demux/mkv: do not expose hidden chapters
Given that the matroska format can contain hidden chapters (and that the
demuxer itself create such where it feels applicable), we should not expose
such chapters (as seekpoints) outside of the demuxer module.

This fixes the ticket 17202 by checking whether or not the
chapter should be displayed to the user, and only appends the item to the
title's seekpoint array if this is the case.

Close #17202

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>