gui/qt: use function that conditionally returns an item in PLModel::isSupported
PLModel::getItem, which was used prior to this commit, will return
either the item associated with an index, or the item referring to the
associated playlist (currently in flight).

Given that we have options that shall only be available if the index is
associated with a given item, simply having a check to see whether
"item" is not null will ease implementation (the alternative is to
compare it to "rootItem", which is redundant).

Signed-off-by: Francois Cartegnie <fcvlcdev@free.fr>