gui/qt: fix PLModel::isSupportedAction( ACTION_PLAY, ... )
Prior to this commit, ACTION_PLAY was supported if:

    - the index was not associated with the currently playing
      item, or;
    - the playlist was paused.

In other words, the option would not be enabled if the playlist was in
a stopped state.

This commit changes the associated logic to the following, meaning
that the action is supported:

    - when the user has right-clicked on an entity, and;
    - the playlist is either stopped or paused, and;
    - the associated index is not currently playing.

Usage of PL_LOCK/PL_UNLOCK has also been replaced by vlc_playlist_lock
to aid code-reasoning and maintenance.

Signed-off-by: Francois Cartegnie <fcvlcdev@free.fr>