gui/qt: clean PLModel::isSupportedAction( ACTION_PAUSE, ... )
There is no need to lock the playlist if the index is not associated
with the current item, as such this commits first checks whether this
condition holds or not, only to then lock the playlist (when it is
actually necessary.

PL_LOCK / PL_UNLOCK has also been replaced by usage of
vlc_playlist_locker.

Signed-off-by: Francois Cartegnie <fcvlcdev@free.fr>