gui/qt: fix usage of rootIndex() in PLModel::isSupportedAction
There is no need to compare the index associated with the right-click,
and the return-value of "rootIndex()" now that VLCModel::getItem is used
instead of PLModel::getItem.

If "item" is NULL, we know that it is not an item within the playlist
(excluding the root).

This commit is also more than justifiable if one looks at the
implementation of PLModel::rootIndex; we should not go through all that
work when a simple NULL-check is sufficient.

Signed-off-by: Francois Cartegnie <fcvlcdev@free.fr>