gui/qt: fix PLModel::isSupportedAction( ACTION_SAVETOPLAYLIST, ... )
Given that the relevant action saves the current playlist (ie.
ROOTTYPE_CURRENT_PLAYING) to a file, it only make sense to show this
option if the user is actually looking at that playlist.

This commit changes the logic associated so that it appears when that is
the case, not just because we have items in the current view.

Signed-off-by: Francois Cartegnie <fcvlcdev@free.fr>