input/demux: make helper type more generic
The affected helper-type can be used elsewhere in this translation-unit though
for lookup of other strings not directly related to a certain mime-type.

This commit changes both the name of the type, but also its data-members, so
that it is more generic. As the lookup maps will always be initialized with
string-literals, there is no need to specify an explicit size for the
data-members; instead we can simply rely on a pointer-to-char.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>