Fix parsing for ATSC channels.conf format
I don't know if there is some channels.conf variation that specified
"VSB_8" as the modulation, but the official dvb-apps scan/azap tool
has used "8VSB" for as long as I can remember (at least going back
to 2008).

Add support for the 8VSB modulation type so that playlists for
ATSC channels.conf work again, but leave in there the existing
"VSB_8" value since it doesn't hurt anything and there might be
some variant of the channels.conf format that I am unfamiliar with.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>