converter: spdif: improve header parser
In function of the codec (like EAC3 in a future commit), we may have to gather
a certain numbers of blocks.

The new parser function returns i_out_size_padded since the size of a S/PDIF
block is not necessarily nb_samples * 4. It also returns i_length_mul since the
length in the S/PDIF header can be in bits or in bytes in function of the
codec.