mkv: fixed memory-leak related to mkv_track_t::fmt
If we do not insert the given track into our maps of tracks, we would
previously not invoke es_format_Clean on the given fmt, which of
course would leak dynamically allocated members.

There was also an issue with assigning values to data-members that
would later be overwritten, of course directly causing dynamically
allocated data to be leaked.

Signed-off-by: Thomas Guillem <thomas@gllm.fr>