access_output: livehttp: update playlist within seglen-periods
Previously we updated playlist only when we got headers from stream. Now
we update playlist/segments when we reach seglen-duration from start of
the segment. At that point code writes all full sets to segment-file and
starts new file.  When we reach header-block we move currently ongoing
set to queue of full sets.