misc/filter_chain: fixup of commit 2875395
The referred to commit assumed that filter->p_cfg always referred to an object
allocated through config_ChainCreate and that ownership was meant to be passed
with values ending up in the data-member (something which is not true).

These changes moves the clean-up to a more appropriate location, while
also making sure that we actually free the allocated data on error in
filter_chain_AppendFromString.

fixes #17429

Signed-off-by: Thomas Guillem <thomas@gllm.fr>