configure: disable -z defs linker check for FreeBSD.
When linking libvlccore with -z defs, the following happens:

posix/.libs/netconf.o: In function `vlc_getProxyUrl': ./src/posix/netconf.c:75:
undefined reference to `environ'

environ on FreeBSD is defined only for the dynamically linked
executables via csu, and is placed in bss section.  It's not available in
libc, which uses the same extern char **environ to access it.

What's interesting is NetBSD took a more compatible approach, where
environ is available both in libc and in csu.