mpg123: rewrite, handle the case 1 input => x outputs
When the mpeg_audio packetizer is not used, input blocks can contain several
frames. The mpg123_decode_frame function is now called if *pp_block is NULL and
can return an output frame if there is enough data in mpg123 internal input.

The audio format is now updated when mpg123 report a format change.