trivial_mixer: don't drop too many channels
This is the trivial mixer, it won't mix anything but just copy channels in the
right order. This commit try to be more clever when dropping unmatching
channels. When channels don't match, try to match a channel of the same side.
For example, if there is an output ML, but no input ML, try to use RL or L
instead.