Fix the tools bootstrap version check
It was broken when ProtoBuf 3.1 was installed system-wide, for example.

The bootstrap script that outputs which binary tools need to be installed, incorrectly
detects that Protocol Buffers 3.1 is a compatible version with 2.6. This changes the
check on the major version number to require it to be same, not just bigger,
since changes on the major version numbers might be breaking changes, as was the case
here.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>