trivial: don't duplicate channels
This is a partial revert of 08d95b393f6471fd31d11ebd9bdbbf5c2dde9a7f

We don't want to copy Stereo chans to Rear and/or Middle chans, even if these
chans won't be used. We should just use Middle chans as Rear chans (if they are
not used) and vice-versa.