contrib: use pkg from contribs before the user ones
If the user directories contain an older or otherwise inadequate
package version, contrib will rebuild a suitable version. If it is a
dependency of another contrib package, then pkg-config needs to look up
the contrib version before the user version.

This behaviour is already what happens for system packages in the
default pkg-config directories: they are processed after the
directories from the environment variable.

(Rewritten description)
Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>