test/url: make it easier to check invalid URLs
The previous implementation, somewhat unconditionally, made
assumptions about the internal state of the result of `vlc_UrlParse`,
meaning that it was inherently difficult to check if invalid URLs was
correctly rejected.

These changes simply asserts that vlc_UrlParse failed if all of the
arguments (used to verify the result) are NULL or 0.

Signed-off-by: Thomas Guillem <thomas@gllm.fr>