stream: rename vlc_stream_NewMRL to vlc_stream_NewURL
The function does not open MRLs, as correctly described in its
documentation, as such it is rather unfortunate that its name uses MRL
instead of URL (especially given that it cannot handle MRLs).

These changes are simply renaming all occurrences of the function, so
that the behavior of the function is properly reflected by its name.

Signed-off-by: Thomas Guillem <thomas@gllm.fr>