demux/mkv: fix 17572: error on cluster without timecode
As can be read in the MKV specification, a cluster must include a
timecode element in order to be valid.

These changes make sure that we error if such element is missing,
effectivelly preventing us from calling KaxCluster::GlobalTimecode
(which will assert if KaxCluster::InitTimecode has not been called).

fixes #17572

Signed-off-by: Steve Lhomme <robux4@videolabs.io>
Signed-off-by: Thomas Guillem <thomas@gllm.fr>