demux/mkv: fix 17575: ignore simple blocks preceeding timecode
As detailed in the matroska specification, the Timecode element within
a cluster is not only mandatory, it shall be the first element within
its parent (given that a SimpleBlock, and other entities, cannot be
interpreted correctly without it).

The previous implementation would assert inside libmatroska on
out-of-order blocks, causing us to crash. These changes are made so
that out-of-order elements are ignored instead of being a
point-of-failure.

An input where the added diagnostic is triggered is ill-formed.

 - https://www.matroska.org/technical/order/index.html

Signed-off-by: Steve Lhomme <robux4@videolabs.io>
Signed-off-by: Thomas Guillem <thomas@gllm.fr>