demux/avi: do not call FromACP with NULL (fixes #17581)
FromACP is used to convert a c-style string from one charset to the
other, as such it does not make sense to call the function with NULL
(especially given that NULL is not a valid input for the function).

These changes fixes the checks to see whether or not the function is
applicable or not by properly checking the argument that would-be
passed (instead of just the object that contains it).

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>