demux/asf: fix 17601: fix undesired integer underflow
This commit shall be viewed as a fix-up of 86835f9.

The previous commit in question did not take into account that the
left-hand side expression can of course result in a negative value,
meaning that the value-promotion necessary for the comparison to take
place would result in a very large value (where we expect a very small
one).

fixes #17601

Signed-off-by: Thomas Guillem <thomas@gllm.fr>