demux/nsv: ignore unsupported blocks
The previous implementation would crash given that p_sys->p_video and
p_sys->p_audio is not guaranteed to be non-NULL (they can be NULL due
to an unsupported codec).

These changes simply make sure that we do not try to send blocks that
do not have have a corresponding ES (causing a crash).

fixes #17571

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>