lua/libs/input: fix 17611: properly populate filename attribute
The previous implementations usage of input_item_GetName resulted in
unexpected data in the filename attribute (given that
input_item_GetName will first query vlc_meta_Title, and if present
return that).

These changes extracts the filename for a given item, including
support for trailing slashes (so that we do not get an empty filename
for a path such as file:///media/).

fixes #17611

Signed-off-by: Pierre Ynard <linkfanel@yahoo.fr>