opengl: fix possible typo
There isn't a case where program[0] is 0 and program[1] is 0. They are always
created together. So the second part of the if() is never called.

program[0] is for YUV and XYZ sources so it is odd to use it with a single
planar texture, given that's what program[1] is for.

Signed-off-by: Thomas Guillem <thomas@gllm.fr>