hotkeys: listen to new inputs and new vouts
Listen to playlist "input-current" to update the current input_thead_t, and
listen to input "intf-event" to update the current vout. The vout/input don't
need to be fetched when processing an action, this removes one FIXME, but this
adds more lock complexity.

This will be mainly needed for the next commit, in order to listen to mouse
events from the vout.