demux/mp4: adjust to changes introduced by 8f9b76e
ExtractIntlStrings is only applicable for a limited range of boxes,
though with the recent changes following its introduction it might be
invoked for entities it is currently not protected from misdiagnosing.

These changes make sure that we only interpret a box inside the
function if:

 - the atom is known as a metadata box
 - the box has been parsed
 - the box has a father, and it is ATOM_udta

The check so that we do not parse boxes that contain a "data"-box has
been removed given that this check is more suitable at where the call
to the function is made.

Signed-off-by: Francois Cartegnie <fcvlcdev@free.fr>