input: fix resuming from pause at EOF (fixes #6490)
As things stood, the input was paused at EOF... and when resumed would
get straight back into pause.

This adds a flag to heep track of the occurrence of pause at EOF. If it
occurs a second time in a row, terminate the input thread loop.

If however the user seeks backward and resume, clear the flag and resume
playback normally.