core: add stream_extractor module type
These changes introduces a new module type named stream_extractor. The
added documentation should explain when such module should be used,
but in short it allows for extraction of data within a stream,
effectively resulting in a new stream that refers to the extracted
data.

Interaction with the stream-extractor shall never happen directly,
instead the module-backend is written in a way so that it exposes a
stream_t to the public.

    [ access ] -> [ stream_t ] -> [ stream consumer ]
                   '- [ stream extractor ]

Future changes are necessary in order to make modules of this type
usable in practice, but has been split into individual commits so that
the changes are easier to follow.

Signed-off-by: Thomas Guillem <thomas@gllm.fr>