playlist/preparser: do not "download" local art
The previous implementation would consider FILE:///art.jpg to be a
remote resources, meaning that it would try to "download" the file and
put it in the cache.

Given that the scheme part of an URI is case-insensitive (as stated in
RFC3986 3.1.), these changes make sure that we treat the affected
schemes in the way which was intended.

fixes #17759

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>