skins2: Remove macOS code
Skins2 is not built on macOS since very long time, given that Qt is not
used there. The code is very old with a lot of TODOs and additionally it
requires the quite old Carbon framework which implies 32bit only as
Carbon was never ported to 64bit. (Cocoa should be used instead)

Therefore this commit removes all macOS related skins2 code.