Use QT file browser functions which return URLs when possible.
The getOpenFileNames() and getSaveFileName() methods of QFileDialog only work
on local paths and return a blank string if the dialog is used to select a
remote file (e.g., on a Samba share).  As of Qt 5.2 the QFileDialog class
provides the methods getOpenFileUrls() and getSaveFileUrl() that return QUrl
objects which can contain URLs to remote paths.  Use these methods when an
appropriate version of the Qt libraries are available so that the paths to
remote files are returned correctly when selected via the various file
selection dialogs.

Signed-off-by: Frank Praznik <frank.praznik@gmail.com>
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>