demux/ttml: remove legacy return values
Previously tt_node_AttributesToText and tt_node_ToText would return a
boolean indicating whether the allocation for the data stream was
successful or not.

Since allocation failures are rare, and that vlc_memstream properly
handles the error state of the stream used, these changes remove the
return-values (as they are no longer needed).

Signed-off-by: Francois Cartegnie <fcvlcdev@free.fr>