macOS: Avoid double-seeking to the same position
The sliders for seeking are continuous, that means they send events
for every action, which is useful to seek when the slider is dragged.
But when the slider is clicked, this results in two seeks nearly at the
same time to the same position.
Therefore this commit introduces a check that ignores the slider update
if it originated from a NSLeftMouseUp event to workaround that.

Ref. #17954