videotoolbox: always do direct rendering
Remove the videotoolbox-zero-copy variable since direct rendering is always
enabled.

Map the CVPX pixel format to a VLC fourcc, this buffer will be renderered by
the OpenGL vout with the cvpx converter (mapping an IOSurface to an OpenGL
texture).

On Macos, the destination pixel format shouldn't set touched at all to let the
decoder output frames in its native format (and don't do any color conversion).
This need to be tested on iOS (hence the #ifdef).