stream_extractor: enable cache for block based extractors
In order to extract an entity from a stream, some heavy lifting is normally
required, as such it makes sense to cache data coming out from a
stream_extractor in order to save CPU (and more importantly battery life on
mobile devices).

This will especially make a big impact when block-based stream_extractors are
used, since they will not be asked to potentially regenerate the same block
data over and over.