macOS: Restructure images a bit
This renames the very generic-named icons folder to better match
what it actually contains, File Icons.
App Icons are moved to their own folder as they are not file icons.
This matches the new structure in Xcode better too.
Pref icons folder is renamed to match other folder names and the icons
are renamed to have a simpler name.