demux/mkv: fix ES handling during DEMUX_SET_TITLE
As part of seeking to a new virtual chapter, we have to destroy the
elementary streams that are part of the current one, and then create
the streams associated with the seeking-to virtual chapter.

This is done in virtual_segment_c::Seek, and is based on whether
p_current_vsegment->p_current_vchapter is equal to the virtual chapter
which we are seeking to.

The previous implementation would set "p_sys->p_current_vsegment->p_current_vchapter"
prior to calling virtual_segment_c::Seek, which makes the
seek-implementation think that we are seeking within the same virtual
chapter (retaining the previous elementary streams, and not creating
new ones).

By removing the relevant assignment under DEMUX_SET_TITLE we let
virtual_segment_c::Seek handle the necessary switching, preventing the
new title to be played from not having the necessary elementary-streams
created (to actually be playable).

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>