contrib: use git-archive and save the hash
git-archive is supposed to validate the git repository to match the
git hash, and the git hash itself is supposed to be cryptographically
secure. So we if save it, we can reuse it to check the tarball.

Note: Alternatively, we could rely on git-archive making reproducible
tarballs, which it indeed does, then perform a regular hash of the
decompressed tarball. It is however unclear if git-archive tarballs are
reproducible *across* Git versions, which could lead to problems.

Note 2: This still requires a head to fetch from. In principles,
fetching a hash directly is possible, but not all servers allow it.